+++ ModulesPath/plotting/__init__.py
@@ -1,3 +1,4 @@
 from .spikes import *
 from .position import *
 from .epochs import *
+from .signals import *
